pyproject.toml
[project]
name = "prop-teststand"
version = "1.0"
authors = [
    {name = "Noah Warren", email = "noahwarren42@gmail.com"}
]
requires-python = ">=3.11"

dependencies = [
    # Running requirements
    "colorama",
    "labjack-ljm",
    "numpy",
    "pyqtgraph",
    "PySide6",
    "PySide6-stubs",
    "keyboard",
    "types-keyboard",

    # Dev stuff
    "ruff",
    "mypy",
    "mpremote",
]

[project.scripts]
# CLI TOOLS
hello_cli = "qretproptools.cli.hello_cli.hello_cli:main"
find_devices = "qretproptools.cli.find_devices.find_devices:main"
direct_tcp = "qretproptools.cli.direct_tcp.direct_tcp:main"

[project.gui-scripts]
# GUI TOOLS
hello_gui = "qretproptools.gui.hello_gui.hello_gui:main"
full_gui =  "qretproptools.gui.full_Gui.full_Gui:main"
multi_plot = "qretproptools.gui.full_Gui.test_multi_plot:main"

[tool.setuptools.packages.find]
where = [""]
include = ["libqretprop*", "qretproptools*"]
namespaces = false

[tool.mypy]
strict = true

[tool.mypy-plotly]
ignore_missing_imports = true

[tool.pyright]
#reportMissingTypeStubs=false   # pyqtgraph has issues
reportPrivateUsage=false       # Leave this to ruff
reportUnknownMemberType=false  # WAY too many problems with PySide6 and pyqtgraph to have this on
reportMissingTypeStubs = "information"